--- Codechef Iem Kolkata chapter- Intermediate/A.cpp
@@ -1,22 +1,0 @@
-#include<bits/stdc++.h>
-#define int long long
-using namespace std;
-
-signed main() {
-
-    int x, ans;
-
-    cin >> x;
-    if (x <= 5)
-        cout << 1 << "\n";
-    else {
-
-        if (x % 5 == 0)
-            cout << x / 5 << "\n";
-        else
-            cout << (x / 5 + 1) << "\n";
-    }
-
-    return 0;
-}
-
