+++ Codechef Iem Kolkata chapter- Intermediate/E/E.py
@@ -1,0 +1,3 @@
+for _ in range(int(input())):
+    k,x=map(int,input().split())
+    print(9*(k-1)+x)
