+++ README.md
@@ -1,2 +1,4 @@
 # Contest-Solutions
 Solutions of the Contests organised by IEM Kolkata Students' Chapter
+
+[Contest Link](https://codeforces.com/gym/306623)
